requirements.txt
pymupdf==1.26.1
requests==2.32.4
python-dotenv==1.1.0
google-genai==1.21.1
pydantic==2.11.7